fixed how app was storing and retrieving info, given dynamodb was set up, but app was still using room.
